Fix distribution
