+++ app/Insightly/Listeners/CreateProjectWithOrganization.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace App\Insightly\Listeners;
 
 use App\Domain\Integrations\Events\IntegrationActivatedWithOrganization;
-use App\Domain\Integrations\Jobs\DistributeKeys;
 use App\Domain\Organizations\Organization;
 use App\Domain\Organizations\Repositories\OrganizationRepository;
 use App\Insightly\InsightlyClient;
@@ -66,7 +65,6 @@ final class CreateProjectWithOrganization implements ShouldQueue
             $organizationInsightlyId
         );
 
-        DistributeKeys::dispatch($integrationActivatedWithOrganization->id);
         $this->logger->info(
             'Project created for integration activated for with',
             [
